Refactoring
Note: This commit has been manually tested by:
1. Enabling VoiceOver (MacOS).
2. Moving VoiceOver focus into the iOS simulator.
3. Opening a note.
4. Opening the note actions menu.
5. Navigating through the options in the menu, verifying that focus does not visit the note title or other parts of the app.
   - An exception to this is the "edit" button, which is shown using a React Native <Portal>, which displays content outside of the region with disabled focus.
6. Moving focus back to the "Attach" button.
7. Pressing the button.
8. Selecting "Record audio" in the attach menu.
9. Verifying that the "Voice recorder" heading is auto-focused.
