+++ .gitignore
@@ -661,10 +661,11 @@ packages/app-mobile/components/TextInput.js
 packages/app-mobile/components/ToggleSpaceButton.js
 packages/app-mobile/components/accessibility/AccessibleModalMenu.js
 packages/app-mobile/components/accessibility/AccessibleView.js
-packages/app-mobile/components/accessibility/FocusControlProvider/AutoFocusProvider.js
-packages/app-mobile/components/accessibility/FocusControlProvider/DialogBlocksAccessibilityFocus.js
-packages/app-mobile/components/accessibility/FocusControlProvider/FocusControlProvider.js
-packages/app-mobile/components/accessibility/FocusControlProvider/ModalContent.js
+packages/app-mobile/components/accessibility/FocusControl/AutoFocusProvider.js
+packages/app-mobile/components/accessibility/FocusControl/FocusControl.js
+packages/app-mobile/components/accessibility/FocusControl/FocusControlProvider.js
+packages/app-mobile/components/accessibility/FocusControl/MainAppContent.js
+packages/app-mobile/components/accessibility/FocusControl/ModalContent.js
 packages/app-mobile/components/app-nav.js
 packages/app-mobile/components/base-screen.js
 packages/app-mobile/components/biometrics/BiometricPopup.js
