+++ packages/app-mobile/components/accessibility/FocusControl/FocusControl.tsx
@@ -1,0 +1,11 @@
+import FocusControlProvider from './FocusControlProvider';
+import MainAppContent from './MainAppContent';
+import ModalContent from './ModalContent';
+
+const FocusControl = {
+	Provider: FocusControlProvider,
+	ModalContent,
+	MainAppContent,
+};
+
+export default FocusControl;
