+++ packages/app-mobile/components/testing/TestProviderStack.tsx
@@ -4,7 +4,7 @@ import { MenuProvider } from 'react-native-popup-menu';
 import { Provider } from 'react-redux';
 import { Store } from 'redux';
 import { AppState } from '../../utils/types';
-import FocusControlProvider from '../accessibility/FocusControlProvider/FocusControlProvider';
+import FocusControl from '../accessibility/FocusControl/FocusControl';
 
 interface Props {
 	store: Store<AppState>;
@@ -13,13 +13,13 @@ interface Props {
 
 const TestProviderStack: React.FC<Props> = props => {
 	return <Provider store={props.store}>
-		<FocusControlProvider>
+		<FocusControl.Provider>
 			<MenuProvider>
 				<PaperProvider>
 					{props.children}
 				</PaperProvider>
 			</MenuProvider>
-		</FocusControlProvider>
+		</FocusControl.Provider>
 	</Provider>;
 };
